pyproject.toml
[tool.poetry]
name = "ai-42-vienna"
version = "0.1.0"
description = ""
authors = ["David Schaupp <ds@umdaschgroup-ventures.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.5"
ultralytics = "^8.0.203"
timm = "^0.9.8"
python-dotenv = "^1.0.0"
streamlit = "^1.28.1"
langchain = "^0.0.329"
beautifulsoup4 = "^4.12.2"
openai = "^0.28.1"
chromadb = "^0.4.15"
tiktoken = "^0.5.1"
google-search-results = "^2.4.2"
duckduckgo-search = "^3.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pose = "ai_42_vienna.vision.pose_estimation:main"
depth = "ai_42_vienna.vision.depth_estimation:main"
object = "ai_42_vienna.vision.object_detection:main"
combined = "ai_42_vienna.vision.combined:main"
chat = "ai_42_vienna.llm.helper:start_chat"
extract = "ai_42_vienna.llm.helper:start_extract"
agent = "ai_42_vienna.llm.helper:start_agent"